Initial database migration and models
Created the models and initial database migration for the project. As
this was being created, I experimented with using Public IDs and only
exporting those "outside the wire." That, however, proved to be a bit
more complicated than I would have liked. I might return to the matter
later. For now, only the integer public IDs are used.